Move controllers to the search package
