+++ kifi-backend/modules/search/app/com/keepit/search/controllers/search/SearchController.scala
@@ -1,0 +1,209 @@
+package com.keepit.search.controllers.search
+
+import com.google.inject.Inject
+import com.keepit.common.controller.SearchServiceController
+import com.keepit.common.db.Id
+import com.keepit.model._
+import com.keepit.search._
+import play.api.mvc.Action
+import views.html
+import com.keepit.model.User
+import scala.concurrent.Await
+import scala.concurrent.duration._
+import play.api.libs.json._
+import com.keepit.search.sharding.ShardSpecParser
+import com.keepit.search.index.user.UserSearchRequest
+import com.keepit.commanders.RemoteUserExperimentCommander
+import play.api.libs.concurrent.Execution.Implicits.defaultContext
+import com.keepit.common.routes.Search
+import com.keepit.search.augmentation.{ AugmentableItem, ItemAugmentationRequest, AugmentationCommander }
+import com.keepit.search.index.graph.library.{ LibraryIndexable, LibraryIndexer }
+import com.keepit.search.controllers.util.SearchControllerUtil
+
+class SearchController @Inject() (
+    searchCommander: SearchCommander,
+    augmentationCommander: AugmentationCommander,
+    languageCommander: LanguageCommander,
+    librarySearchCommander: LibrarySearchCommander,
+    userSearchCommander: UserSearchCommander,
+    distributedSearchClient: DistributedSearchServiceClient,
+    userExperimentCommander: RemoteUserExperimentCommander) extends SearchServiceController {
+
+  def distSearch() = Action(parse.tolerantJson) { request =>
+    val json = request.body
+    val shardSpec = (json \ "shards").as[String]
+    val searchRequest = (json \ "request")
+
+    // keep the following in sync with SearchServiceClientImpl
+    val userId = (searchRequest \ "userId").as[Long]
+    val lang1 = (searchRequest \ "lang1").as[String]
+    val lang2 = (searchRequest \ "lang2").asOpt[String]
+    val query = (searchRequest \ "query").as[String]
+    val filter = (searchRequest \ "filter").asOpt[String]
+    val maxHits = (searchRequest \ "maxHits").as[Int]
+    val context = (searchRequest \ "context").asOpt[String]
+    val debug = (searchRequest \ "debug").asOpt[String]
+
+    val id = Id[User](userId)
+    val userExperiments = Await.result(userExperimentCommander.getExperimentsByUser(id), 5 seconds)
+    val shards = (new ShardSpecParser).parse[NormalizedURI](shardSpec)
+    val result = searchCommander.distSearch(
+      shards,
+      id,
+      Lang(lang1),
+      lang2.map(Lang(_)),
+      userExperiments,
+      query,
+      filter,
+      maxHits,
+      context,
+      None,
+      debug)
+
+    Ok(result.json)
+  }
+
+  def distSearch2() = Action.async(parse.tolerantJson) { request =>
+    val json = request.body
+    val shardSpec = (json \ "shards").as[String]
+    val searchRequest = (json \ "request")
+
+    // keep the following in sync with SearchServiceClientImpl
+    val userId = (searchRequest \ "userId").as[Long]
+    val lang1 = (searchRequest \ "lang1").as[String]
+    val lang2 = (searchRequest \ "lang2").asOpt[String]
+    val query = (searchRequest \ "query").as[String]
+    val filter = (searchRequest \ "filter").asOpt[String]
+    val libraryContext = ((searchRequest \ "authorizedLibrary").asOpt[Long], (searchRequest \ "library").asOpt[Long]) match {
+      case (Some(libId), _) => LibraryContext.Authorized(libId)
+      case (None, Some(libId)) => LibraryContext.NotAuthorized(libId)
+      case _ => LibraryContext.None
+    }
+    val maxHits = (searchRequest \ "maxHits").as[Int]
+    val context = (searchRequest \ "context").asOpt[String]
+    val debug = (searchRequest \ "debug").asOpt[String]
+
+    val id = Id[User](userId)
+    val userExperiments = Await.result(userExperimentCommander.getExperimentsByUser(id), 5 seconds)
+    val shards = (new ShardSpecParser).parse[NormalizedURI](shardSpec)
+    searchCommander.distSearch2(
+      shards,
+      id,
+      Lang(lang1),
+      lang2.map(Lang(_)),
+      userExperiments,
+      query,
+      filter,
+      libraryContext,
+      maxHits,
+      context,
+      None,
+      debug).map { result => Ok(result.json) }
+  }
+
+  def distLangFreqs() = Action.async(parse.tolerantJson) { request =>
+    val json = request.body
+    val shardSpec = (json \ "shards").as[String]
+    val searchRequest = (json \ "request")
+
+    val userId = Id[User]((searchRequest \ "userId").as[Long])
+    val libraryContext = ((searchRequest \ "authorizedLibrary").asOpt[Long], (searchRequest \ "library").asOpt[Long]) match {
+      case (Some(libId), _) => LibraryContext.Authorized(libId)
+      case (None, Some(libId)) => LibraryContext.NotAuthorized(libId)
+      case _ => LibraryContext.None
+    }
+    val shards = (new ShardSpecParser).parse[NormalizedURI](shardSpec)
+    languageCommander.distLangFreqs(shards, userId, libraryContext).map { freqs =>
+      Ok(Json.toJson(freqs.map { case (lang, freq) => lang.lang -> freq }))
+    }
+  }
+
+  def distAugmentation() = Action.async(parse.tolerantJson) { request =>
+    val json = request.body
+    val shardSpec = (json \ "shards").as[String]
+    val shards = (new ShardSpecParser).parse[NormalizedURI](shardSpec)
+    val augmentationRequest = (json \ "request").as[ItemAugmentationRequest]
+    augmentationCommander.distAugmentation(shards, augmentationRequest).map { augmentationResponse =>
+      Ok(Json.toJson(augmentationResponse))
+    }
+  }
+
+  def distLibrarySearch() = Action.async(parse.tolerantJson) { request =>
+    val json = request.body
+    val shardSpec = (json \ "shards").as[String]
+    val shards = (new ShardSpecParser).parse[NormalizedURI](shardSpec)
+    val librarySearchRequest = (json \ "request").as[LibrarySearchRequest]
+    librarySearchCommander.distLibrarySearch(shards, librarySearchRequest).map { libraryShardResults =>
+      Ok(Json.toJson(libraryShardResults))
+    }
+  }
+
+  //internal (from eliza/shoebox)
+  def warmUpUser(userId: Id[User]) = Action { request =>
+    searchCommander.warmUp(userId)
+    Ok
+  }
+
+  def searchWithConfig() = Action(parse.tolerantJson) { request =>
+    val js = request.body
+    val userId = Id[User]((js \ "userId").as[Long])
+    val query = (js \ "query").as[String]
+    val maxHits = (js \ "maxHits").as[Int]
+    val predefinedConfig = (js \ "config").as[Map[String, String]]
+    val res = searchCommander.search(userId, acceptLangs = Seq(), experiments = Set.empty, query = query, filter = None, maxHits = maxHits, lastUUIDStr = None, context = None, predefinedConfig = Some(SearchConfig(predefinedConfig)))
+    Ok(JsArray(res.hits.map { x =>
+      val id = x.uriId.id
+      val title = x.bookmark.title.getOrElse("")
+      val url = x.bookmark.url
+      Json.obj("uriId" -> id, "title" -> title, "url" -> url)
+    }))
+  }
+
+  def searchUsers() = Action(parse.tolerantJson) { request =>
+    val userSearchRequest = Json.fromJson[UserSearchRequest](request.body).get
+    val res = userSearchCommander.searchUsers(userSearchRequest)
+    Ok(Json.toJson(res))
+  }
+
+  def userTypeahead() = Action(parse.json) { request =>
+    val userSearchRequest = Json.fromJson[UserSearchRequest](request.body).get
+    val res = userSearchCommander.userTypeahead(userSearchRequest, excludedExperiments = Seq("fake")) // TODO(yingjie): Address admins differently
+    Ok(Json.toJson(res))
+  }
+
+  def explain(query: String, userId: Id[User], uriId: Id[NormalizedURI], lang: Option[String], debug: Option[String]) = Action.async { request =>
+    if (searchCommander.findShard(uriId).isDefined) {
+      val userExperiments = Await.result(userExperimentCommander.getExperimentsByUser(userId), 5 seconds)
+      searchCommander.explain(userId, uriId, lang, userExperiments, query, debug) map { explanationOpt =>
+        Ok(html.admin.explainResult(userId, uriId, explanationOpt))
+      }
+    } else {
+      distributedSearchClient.call(userId, uriId, Search.internal.explain(query, userId, uriId, lang, debug), JsNull).map(r => Ok(r.body))
+    }
+  }
+
+  def augmentation() = Action.async(parse.json) { implicit request =>
+    augmentationCommander.augmentation(request.body.as[ItemAugmentationRequest]).map { augmentationResponse =>
+      val json = Json.toJson(augmentationResponse)
+      Ok(json)
+    }
+  }
+
+  def augment() = Action.async(parse.json) { implicit request =>
+    // This should stay in sync with SearchServiceClient.augment
+    val userId = (request.body \ "userId").asOpt[Id[User]]
+    val maxKeepersShown = (request.body \ "maxKeepersShown").as[Int]
+    val maxLibrariesShown = (request.body \ "maxLibrariesShown").as[Int]
+    val maxTagsShown = (request.body \ "maxTagsShown").as[Int]
+    val items = (request.body \ "items").as[Seq[AugmentableItem]]
+    val showPublishedLibrariesOpt = (request.body \ "showPublishedLibraries").asOpt[Boolean]
+
+    val itemAugmentationRequest = ItemAugmentationRequest.uniform(userId getOrElse SearchControllerUtil.nonUser, items: _*).copy(showPublishedLibraries = showPublishedLibrariesOpt)
+    augmentationCommander.getAugmentedItems(itemAugmentationRequest).map { augmentedItems =>
+      val infos = items.map(augmentedItems(_).toLimitedAugmentationInfo(maxKeepersShown, maxLibrariesShown, maxTagsShown))
+      val result = Json.toJson(infos)
+      Ok(result)
+    }
+  }
+}
+
