+++ kifi-backend/modules/search/test/com/keepit/search/common/store/SearchFakeStoreModule.scala
